ecocrypt/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
librosa==0.10.1
numpy==1.24.3
soundfile==0.12.1
pycryptodome==3.18.0
pymongo==4.5.0
python-dotenv==1.0.0